# gcp/api_gateway/openapi-functions.yaml
swagger: '2.0'
info:
  title: myapi
  description: Sample API on API Gateway with a Google Cloud Functions backend
  version: 1.0.0
host: myapi.apigateway.gcpsteamanalytics.cloud.goog
schemes:
  - https
produces:
  - application/json
paths:
...
  /getReviews/{appid}/{review_type}/{limit}:
    get:
      summary: Get reviews by app ID, type, and limit
      operationId: getReviews
      x-google-backend:
        address: https://us-central1-gcpsteamanalytics.cloudfunctions.net/get-reviews
        path_translation: APPEND_PATH_TO_ADDRESS
      parameters:
        - name: appid
          in: path
          required: true
          type: string
        - name: review_type
          in: path
          required: true
          type: string
        - name: limit
          in: path
          required: true
          type: integer
      responses:
        '200':
          description: A successful response
          schema:
            type: string